progress on jsonrpc tests
