+++ external-crates/move/crates/move-package-alt/tests/data/pinning/external_basic/Move@pinned.snap.new
@@ -1,0 +1,5 @@
+---
+source: crates/move-package-alt/tests/pinning_tests.rs
+assertion_line: 36
+---
+mock-resolver did not follow the external resolver protocol: unknown field `result` at line 1 column 58
